Update Gogoanime and Fix download resume issues (#197)
* Rely on AnimePahe for episode naming

* Remove use of enumerate

* Add useful debug info for mp4upload

* Fix minor regex mishap for mp4upload

* Better title naming for mp4upload

* Minor tweaks complete

* MP4Upload regex minor improvement

* Make collection of sources look better

* Revert back to using enumerate for episode numbering

* Added utility function to parse episode range

* Replace episode range collecting with utility function to parse episode range

* Add grammar option to cli.py

* Make grammar more consistent

* Implement grammar parser and add as util function

* Added search to gogoanime

* Enable getting episode sources for Gogoanime

* Minor refactor for grammar parser

* Use new episode parser by default and add gogoanime to provider choices

* Fix minor oversight to identify None type passed to episode parser

* Remove explicit checks for None type in episode string parsers

* Enable retries for request session

* Make cfscrape capable of retrying

* Make provider list more readable in cli.py

* Handle failure to find stream URL better in MP4Upload extractor

* Revert changes to match master

* Update gogoanime domain

* Fix failure to skip already downloaded files

* Fix potential bug

* Enable ranged download to resume stopped download

* Avoid constantly opening and closing file in downloader

+++ anime_downloader/downloader/http_downloader.py
@@ -1,4 +1,5 @@
 import os
+import logging
 
 from anime_downloader.downloader.base_downloader import BaseDownloader
 from anime_downloader import session
@@ -8,39 +9,60 @@ session = session.get_session()
 
 class HTTPDownloader(BaseDownloader):
     def _download(self):
+        if os.path.exists(self.path):
+            # Handle both cases where user will still want file downloaded
+            # even if it exists
+            if abs(os.stat(self.path).st_size - self.total_size) < 10 and not self.force:
+                logging.warning('File already downloaded. Skipping download.')
+                return
+            elif self.force:
+                os.remove(self.path)
+
         if self.range_size is None:
             self._non_range_download()
         else:
             self._ranged_download()
 
+    def pre_process(self):
+        self.downloaded = 0
+        # Preprocess ranges and amount downloaded for ranged download
+        if self.range_size:
+            if not os.path.exists(self.path):
+                self._range_start = 0
+                self.downloaded = 0
+                self._range_end = self.range_size
+
+                # Make a new file, maybe not the best way
+                with open(self.path, 'w'):
+                    pass
+            else:
+                self._range_start = os.stat(self.path).st_size
+                self._range_end = self._range_start + self.range_size
+                self.downloaded = self._range_start
+
     def _ranged_download(self):
         http_chunksize = self.range_size
 
-        range_start = 0
-        range_end = http_chunksize
+        range_start = self._range_start
+        range_end = self._range_end
 
-        # Make a new file, maybe not the best way
-        with open(self.path, 'w'):
-            pass
+        with open(self.path, 'ab') as f:
+            while self.downloaded < self.total_size:
+                r = session.get(self.url,
+                                headers=set_range(range_start, range_end, self.referer),
+                                stream=True)
+                if r.status_code == 206:
+                        for chunk in r.iter_content(chunk_size=self.chunksize):
+                            if chunk:
+                                f.write(chunk)
+                                self.report_chunk_downloaded()
 
-        r = session.get(self.url, headers={'referer': self.referer}, stream=True)
-        while self.downloaded < self.total_size:
-            r = session.get(self.url,
-                            headers=set_range(range_start, range_end, self.referer),
-                            stream=True)
-            if r.status_code == 206:
-                with open(self.path, 'ab') as f:
-                    for chunk in r.iter_content(chunk_size=self.chunksize):
-                        if chunk:
-                            f.write(chunk)
-                            self.report_chunk_downloaded()
-
-            if range_end == '':
-                break
-            range_start = os.stat(self.path).st_size
-            range_end += http_chunksize
-            if range_end > self.total_size:
-                range_end = ''
+                if range_end == '':
+                    break
+                range_start = os.stat(self.path).st_size
+                range_end = range_start + http_chunksize
+                if range_end > self.total_size:
+                    range_end = ''
 
     def _non_range_download(self):
         r = session.get(self.url, headers={'referer': self.referer}, stream=True)
